pyproject.toml
[project]
dynamic = ["dependencies"]
[tool.setuptools.dynamic]
dependencies = {file = ["requirements.txt"]}
[tool.poetry]
name = "llm-slack-bot"
version = "0.1.0"
description = ""
authors = ["Sateesh Kadiyala <skadiyala@inkling.com>"]
readme = "README.md"
packages = [{include = "llm_slack_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.277"
gpt4all = "^1.0.3"
urllib3 = "==2.0.3"
PyMuPDF = "==1.22.5"
unstructured = "==0.8.0"
extract-msg = "==0.41.5"
tabulate = "==0.9.0"
pandoc = "==2.3"
pypandoc = "==1.11"
tqdm = "==4.65.0"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"
fastapi = "^0.103.0"
sentence-transformers = "^2.2.2"
psycopg2 = "^2.9.7"
pgvector = "^0.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
